refactor(thumbnail): move everything to `thumbnail` model
